+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Documentation to profiles.
 - `Redis` service with profile `cache`.
 - Make: Targets for `debug` and `info`.
+- PHP: `docker-entrypoint`
 
 ### Changed